chore(create-novu-app): rename @novu/echo to @novu/framework
Apply PR comments
