+++ packages/create-novu-app/templates/app-react-email/ts/app/novu/emails/vercel.tsx
