+++ packages/create-novu-app/templates/app-react-email/ts/app/novu/workflows.ts
