+++ packages/create-novu-app/templates/app/ts/app/api/novu/route.ts
@@ -1,6 +1,6 @@
 import { serve } from "@novu/framework/next";
 import { Client } from "@novu/framework";
-import { myWorkflow } from "../../workflow/myWorkflow";
+import { myWorkflow } from "../../novu/workflows";
 
 const client = new Client({ strictAuthentication: false });
