+++ packages/create-novu-app/templates/app/ts/app/novu/workflows.ts
